adding Dianas profile pages
