+++ app/routes/profile/profile.tsx
@@ -1,217 +1,34 @@
-import { useOutletContext } from "react-router";
+import { Outlet, useOutletContext } from "react-router";
 import { type AuthData } from "~/services/api";
-import { Navigate } from "react-router";
-import React, { useEffect, useState } from "react";
-
-interface borrowedItems {
-  ItemID: number;
-  Title: string;
-  DueDate?: string;
-}
+import { Link } from "react-router";
+import "./profile.css";
 
 export default function ProfilePage() {
-  const {
-    isLoggedIn,
-    memberID,
-    groupID,
-    firstName,
-    lastName,
-    middleName,
-    address,
-  } = useOutletContext<AuthData>();
-  const [items, setItems] = React.useState<borrowedItems[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [selectedItems, setSelectedItems] = useState<borrowedItems[]>([]);
-  const [returnStatus, setReturnStatus] = useState<string | null>(null);
-  const [isReturning, setIsReturning] = useState(false);
-
-  const fetchData = () => {
-    setIsLoading(true);
-    setError(null);
-    fetch(`/profile/api/borroweditems/${memberID}`)
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-        return response.json();
-      })
-      .then((data) => {
-        setItems(data);
-        setIsLoading(false);
-      })
-      .catch((error) => {
-        setError(`Failed to fetch items: ${error.message}`);
-        setIsLoading(false);
-      });
-  };
-
-  useEffect(() => {
-    fetchData();
-  }, []);
-
-  const handleSelectItem = (item: borrowedItems) => {
-    setSelectedItems((prev) => {
-      // Check if item is already selected
-      const isSelected = prev.some(
-        (selectedItem) => selectedItem.ItemID === item.ItemID
-      );
-
-      if (isSelected) {
-        // Remove item if already selected
-        return prev.filter(
-          (selectedItem) => selectedItem.ItemID !== item.ItemID
-        );
-      } else {
-        // Add item if not selected
-        return [...prev, item];
-      }
-    });
-  };
-
-  const isItemSelected = (itemId: number) => {
-    return selectedItems.some((item) => item.ItemID === itemId);
-  };
-
-  const handleReturnItems = () => {
-    if (selectedItems.length === 0) {
-      setReturnStatus("No items selected for return");
-      return;
-    }
-
-    setIsReturning(true);
-    setReturnStatus("Processing return...");
-
-    const itemIds = selectedItems.map((item) => item.ItemID);
-
-    fetch(`api/return`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        items: itemIds,
-      }),
-    })
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-        return response.json();
-      })
-      .then((data) => {
-        setReturnStatus(
-          `Successfully returned ${selectedItems.length} item(s)`
-        );
-        setSelectedItems([]);
-        // Refresh the list after successful return
-        fetchData();
-        setIsReturning(false);
-      })
-      .catch((error) => {
-        setReturnStatus(`Failed to return items: ${error.message}`);
-        setIsReturning(false);
-      });
-  };
-
-  // if user is not logged in
-  if (!isLoggedIn) {
-    return <Navigate to="/login" />;
-  }
+  const authData = useOutletContext<AuthData>();
 
   return (
     <div>
-      <h2>Welcome to profile page #{memberID}</h2>
-      <h2>{groupID}</h2>
-      <h2>First Name: {firstName}</h2>
-      <h2>Middle Name: {middleName}</h2>
-      <h2>Last Name: {lastName}</h2>
-      <h2>Address: {address}</h2>
-      <h1>Add more fields for birthday, balance, etc.</h1>
-
-      {isLoading ? (
-        <p>Loading borrowed items...</p>
-      ) : error ? (
-        <p style={{ color: "red" }}>{error}</p>
-      ) : items.length > 0 ? (
-        <div>
-          <h3>Borrowed Items</h3>
-
-          {returnStatus && (
-            <div
-              style={{
-                margin: "10px 0",
-                padding: "10px",
-                backgroundColor: returnStatus.includes("Failed")
-                  ? "#ffebee"
-                  : "#e8f5e9",
-                borderRadius: "4px",
-              }}
-            >
-              {returnStatus}
-            </div>
-          )}
-
-          <table style={{ width: "100%", borderCollapse: "collapse" }}>
-            <thead>
-              <tr>
-                <th style={{ textAlign: "left", padding: "8px" }}>Select</th>
-                <th style={{ textAlign: "left", padding: "8px" }}>Title</th>
-                <th style={{ textAlign: "left", padding: "8px" }}>Due Date</th>
-              </tr>
-            </thead>
-            <tbody>
-              {items.map((item) => (
-                <tr
-                  key={item.ItemID}
-                  style={{
-                    backgroundColor: isItemSelected(item.ItemID)
-                      ? "#f0f0ff"
-                      : "transparent",
-                    border: "1px solid #ddd",
-                  }}
-                >
-                  <td style={{ padding: "8px" }}>
-                    <input
-                      type="checkbox"
-                      checked={isItemSelected(item.ItemID)}
-                      onChange={() => handleSelectItem(item)}
-                    />
-                  </td>
-                  <td style={{ padding: "8px" }}>{item.Title}</td>
-                  <td style={{ padding: "8px" }}>
-                    {item.DueDate &&
-                      new Date(item.DueDate).toLocaleDateString()}
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-
-          {selectedItems.length > 0 && (
-            <div style={{ marginTop: "20px" }}>
-              <p>{selectedItems.length} item(s) selected</p>
-              <button
-                onClick={handleReturnItems}
-                disabled={isReturning}
-                style={{
-                  padding: "10px 16px",
-                  backgroundColor: "#4CAF50",
-                  color: "white",
-                  border: "none",
-                  borderRadius: "4px",
-                  cursor: isReturning ? "not-allowed" : "pointer",
-                  opacity: isReturning ? 0.7 : 1,
-                }}
-              >
-                {isReturning ? "Processing..." : "Return Selected Items"}
-              </button>
-            </div>
-          )}
+      <div className="bg-nav3">
+        <div className="navbar-left w-full flex justify-between items-center">
+          <ul className="flex space-x-10">
+            <li className="text-nav3">
+              <Link to="./dashboard">Dashboard</Link>
+            </li>
+            <li className="text-nav3">
+              <Link to="./myitems">My Items</Link>
+            </li>
+            <li className="mailbox-nav3">
+              <Link to="/mailbox">✉️</Link>
+            </li>
+            <li className="settings-nav3">
+              <Link to="/settings">⚙️</Link>
+            </li>
+          </ul>
         </div>
-      ) : (
-        <div>No items checked out</div>
-      )}
+      </div>
+      <main>
+        <Outlet context={authData} />
+      </main>
     </div>
   );
 }
